Pedidos sin comillas

+++ ProyectoRedes/app/src/main/java/com/dgfp/proyectoredes/OrdenarActivity.kt
@@ -131,7 +131,7 @@ class OrdenarActivity : AppCompatActivity() {
         )
 
         val objComida = DCComida(
-            Id_Comida = idComida
+            Id_Comida = idComida.toInt()
         )
 
         val objPostPedido = DCPostPedido(
